further fixing of how A### is displayed

+++ Modules/CustomModes.cs
@@ -51,7 +51,7 @@ namespace DynamicMoonRatings.Modules
             }
             if (Plugin.displayMode == "A###")
             {
-                int modifiedRating = (int)(rating - (1000 * (int)Math.Floor((decimal)(rating / 1000))));
+                int modifiedRating = (int)(rating - (1000 * (int)Math.Floor((float)(rating / 1000))));
                 Plugin.Logger.LogDebug(Plugin.displayMode + " CustomDisplayString Set: " + displayRating + (modifiedRating).ToString("000"));
                 return displayRating + (modifiedRating).ToString("000");
             }
